Rename attendees to networking
