+++ src/app/networking/page.tsx
@@ -2,13 +2,13 @@
 
 import { useState } from "react"
 
-import { ConferenceSelection } from "@/app/attendees/conference-selection"
-import { ProfileList } from "@/app/attendees/profile-list.component"
+import { ConferenceSelection } from "@/app/networking/conference-selection"
+import { ProfileList } from "@/app/networking/profile-list.component"
 import { NoticeText } from "@/components/ui/messages/notice-text"
 import { PageTitle } from "@/components/ui/page-title"
 import { api } from "@/trpc/react"
 
-export default function AttendeesPage() {
+export default function NetworkingPage() {
   const apiFetchAttendingConferences = api.attendance.myConferences.useQuery()
 
   const [selectedConferenceId, setSelectedConferenceId] = useState<string>("")
@@ -16,7 +16,7 @@ export default function AttendeesPage() {
   return (
     <div>
       <PageTitle
-        title={"Attendees"}
+        title={"Networking"}
         subtitle={
           "A list of all the people who have tickets for the selected conference"
         }
@@ -31,7 +31,7 @@ export default function AttendeesPage() {
       {selectedConferenceId ? (
         <ProfileList conferenceId={selectedConferenceId} />
       ) : (
-        <NoticeText text={"Select a conference to view attendees"} />
+        <NoticeText text={"Select a conference to view participants"} />
       )}
     </div>
   )
